# app/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: OpenWatch (Stage 0)
  version: 0.1.0-alpha.1
  description: |
    Stage 0 walking-skeleton surface — three endpoints that exercise the
    foundation stack (correlation, idempotency, audit, error envelope).
servers:
  - url: https://localhost:8443
    description: Local development
paths:
...
  /api/v1/admin/users:
    get:
      operationId: getAdminUsers
      summary: List users
      x-required-permission: user:read
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UsersListResponse'}
        '403':
          description: Caller lacks user:read
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    post:
      operationId: postAdminUsers
      summary: Create a user
      x-required-permission: user:write
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserCreateRequest'}
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserResponse'}
        '400':
          description: Validation or policy failure
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '403':
          description: Caller lacks user:write
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/admin/users/{id}:
    get:
      operationId: getAdminUserByID
      summary: Fetch a user
      x-required-permission: user:read
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: User
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserResponse'}
        '404':
          description: User not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    delete:
      operationId: deleteAdminUserByID
      summary: Soft-delete a user
      x-required-permission: user:delete
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '204':
          description: Deleted
        '404':
          description: User not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/admin/users/{id}/roles:assign:
    post:
      operationId: postAdminUserRolesAssign
      summary: Assign a role to a user
      x-required-permission: role:assign
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserRoleAssignRequest'}
      responses:
        '204':
          description: Role assigned
        '400':
          description: Unknown role id
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/admin/users/{id}/roles:unassign:
    post:
      operationId: postAdminUserRolesUnassign
      summary: Remove a role from a user
      x-required-permission: role:assign
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserRoleAssignRequest'}
      responses:
        '204':
          description: Role removed (idempotent)
  /api/v1/admin/roles:create:
    post:
      operationId: postAdminRolesCreate
      summary: Create a custom role
      x-required-permission: role:write
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CustomRoleCreateRequest'}
      responses:
        '201':
          description: Custom role created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CustomRoleResponse'}
        '400':
          description: Permission unknown
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '409':
          description: Role id taken
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
components:
...
  schemas:
...
    UserResponse:
      type: object
      required: [id, username, email, is_admin]
      properties:
        id:                       {type: string, format: uuid}
        username:                 {type: string}
        email:                    {type: string}
        is_admin:                 {type: boolean}
        last_password_change_at:  {type: string, format: date-time}
        created_at:               {type: string, format: date-time}
        updated_at:               {type: string, format: date-time}
    UsersListResponse:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items: {$ref: '#/components/schemas/UserResponse'}
    UserCreateRequest:
      type: object
      required: [username, email, password]
      properties:
        username: {type: string, minLength: 1, maxLength: 256}
        email:    {type: string, minLength: 3, maxLength: 256}
        password: {type: string, minLength: 8, maxLength: 256}
        is_admin: {type: boolean}
    UserRoleAssignRequest:
      type: object
      required: [role_id]
      properties:
        role_id: {type: string, minLength: 1}
    CustomRoleCreateRequest:
      type: object
      required: [id, description, permissions]
      properties:
        id:          {type: string, pattern: '^[a-z][a-z0-9_]{1,63}$'}
        description: {type: string, minLength: 1, maxLength: 256}
        permissions:
          type: array
          minItems: 1
          items: {type: string}
    CustomRoleResponse:
      type: object
      required: [id, description, is_built_in, permissions]
      properties:
        id:           {type: string}
        description:  {type: string}
        is_built_in:  {type: boolean}
        permissions:
          type: array
          items: {type: string}
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, fault, retryable, human_message]
          properties:
            code: {type: string}
            fault: {type: string, enum: [client, server, policy, external]}
            retryable: {type: boolean}
            human_message: {type: string}
            detail: {type: object, additionalProperties: true}
            correlation_id: {type: string}
...